Update playlists api

+++ app/Http/Controllers/Admin/PlaylistsController.php
@@ -84,11 +84,15 @@ class PlaylistsController extends Controller
      */
     public function update(Request $request, Playlist $playlist)
     {
+        if (Playlist::featured()->exists() && ! $playlist->is_featured)
+            return redirect()->back()->with('error', Playlist::featured()->first()->title . ' is already featured.');
+
         $playlist->update([
             'name' => $request->name,
             'subtitle' => $request->subtitle,
             'group' => $request->group ?? null,
-            'description' => $request->description
+            'description' => $request->description,
+            'featured' => $request->featured ?? null
         ]);
 
         $playlist->pieces()->detach();
